+++ Scripts/GenericProgressBar.cs
@@ -6,7 +6,7 @@ using UnityEngine.UI;
 public class GenericProgressBar : MonoBehaviour
 { 
     public Image bar;  
-    public void SetRatio(int current, float max)
+    public void SetRatio(float current, float max)
     {
         if (bar == null) return;
         float ratio = current / max;
